// frontend/src/components/shared/Text/types.ts
export type TextTag = 'span' | 'p' | 'h1' | 'h2' | 'h3' | 'h4' | 'h5' | 'h6'

export type TextAlign = 'left' | 'right' | 'center'

export type TextColor =
  | 'red'
  | 'blue'
  | 'yellow'
  | 'green'
  | 'orange'
  | 'purple'
  | 'pink'
  | 'black'
  | 'white'
  | 'gray'
  | 'primary'
  | 'secondary'
  | 'heading'

export type TextFontWeight = 'regular' | 'medium' | 'semibold' | 'bold' | number

export type TextSize = 'large' | 'medium' | 'small'